big-league/backend/requirements.txt
amply==0.1.4
aniso8601==7.0.0
asgiref==3.2.3
certifi==2021.10.8
cffi==1.13.2
chardet==3.0.4
config==0.5.0.post0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-extensions==3.1.3
django-filter==2.3.0
django-graphql-jwt==0.3.1
django-heroku==0.3.1
django-model-utils==4.0.0
django-pandas==0.6.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.18
Faker==4.0.0
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==2.8
names==0.3.0
numpy==1.17.3
pandas==1.2.4
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.9.3
PuLP==2.5.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2019.3
requests==2.22.0
Rx==1.6.1
# scipy==1.5.3
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
Unidecode==1.1.1
urllib3==1.25.7
whitenoise==5.3.0
